refactor: put the api in a separate file
